update refactorized code and implemente test

+++ BooksReviews.Application/Features/Books/Commands/DeleteBook/DeleteBookCommand.cs
@@ -1,11 +1,12 @@
 using MediatR;
 using BooksReviews.Application.Common.Interfaces;
+using BooksReviews.Application.Common.Models;
 
 namespace BooksReviews.Application.Features.Books.Commands.DeleteBook;
 
-public record DeleteBookCommand(string Id) : IRequest<Unit>;
+public record DeleteBookCommand(string Id) : IRequest<Result>;
 
-public class DeleteBookCommandHandler : IRequestHandler<DeleteBookCommand, Unit>
+public class DeleteBookCommandHandler : IRequestHandler<DeleteBookCommand, Result>
 {
     private readonly IBookRepository _bookRepository;
 
@@ -14,9 +15,13 @@ public class DeleteBookCommandHandler : IRequestHandler<DeleteBookCommand, Unit>
         _bookRepository = bookRepository;
     }
 
-    public async Task<Unit> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
+    public async Task<Result> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
     {
+        var book = await _bookRepository.GetByIdAsync(request.Id);
+        if (book == null)
+            return Result.Failure("Not Found");
+
         await _bookRepository.DeleteAsync(request.Id);
-        return Unit.Value;
+        return Result.Success();
     }
 }
